refactor(preload): extract renderer log forwarding and the sentry renderer ipc bridge
src/preload/index.ts carries two app-shell concerns that have nothing to do
with building the electronAPI surface: the console.warn/console.error
forwarder that ships prefixed renderer logs over IPC, and the classic Sentry
renderer IPC bridge that keeps packaged renderers off the sentry-ipc:// fetch
fallback. Both are self-contained, and both sit in the middle of the file's
import and helper region.

Move them to two flat siblings:

- src/preload/installRendererLogForwarding.ts exports
  installRendererLogForwarding and keeps formatConsoleArg,
  shouldForwardConsoleText and MAX_FORWARDED_RENDERER_LOG_CHARS module-private.
  RENDERER_LOG moves with it and is dropped from the ipcChannels import in
  index.ts.
- src/preload/installSentryRendererIpcBridge.ts exports
  installSentryRendererIpcBridge and keeps SentryIpcChannel,
  SentryRendererIpcBridge, SENTRY_IPC_NAMESPACE, createSentryIpcKey and the
  Window.__SENTRY_IPC__ global augmentation module-private.

This is pure code motion. The moved bodies are byte-identical apart from the
added export keyword on the two entry points. Both new modules are plain
declarations with no top-level side effects, so preload execution order is
unchanged: the Sentry install try/catch still runs where it did, then
installRendererLogForwarding(), then the RENDERER_BOOT signal, then the
heartbeat interval. No public symbol is renamed and no visibility widens.

src/preload/ previously held only index.ts and constants/ipcChannels.ts, so
there is no in-folder precedent for a flat sibling; every existing
create*Bridge.ts lives under src/features/<feature>/preload/. A flat sibling is
the right home here because these are app-shell concerns rather than
feature-scoped bridges, but reviewers who prefer src/preload/bridges/ can say
so and the move is trivial.

src/preload/index.ts goes from 2094 to 2005 lines; its frozen cap of 2094 is
left alone. scripts/ci/verify-sentry-release.cjs greps the built preload output
for '__SENTRY_IPC__' and 'sentry-ipc'; it still passes because
electron.vite.config.ts has a single preload rollup input and the check walks
every file under dist-electron/preload. Anyone who later externalizes these
modules must re-check that script.

test/preload/electronApiMemberWorkSync.test.ts is unchanged and still passes.
